Move common theme components and manager to courier-ui-core.
