+++ @trycourier/courier-ui-inbox/src/types/courier-inbox-theme-manager.ts
@@ -1,149 +1,55 @@
-import { CourierComponentThemeMode, SystemThemeMode, addSystemThemeModeListener, getSystemThemeMode } from "@trycourier/courier-ui-core";
+import { CourierThemeManager, CourierThemeSubscription, SystemThemeMode } from "@trycourier/courier-ui-core";
 import { CourierInboxTheme, defaultDarkTheme, defaultLightTheme, mergeTheme } from "./courier-inbox-theme";
 
-export interface CourierInboxThemeSubscription {
+export interface CourierInboxThemeSubscription extends CourierThemeSubscription<CourierInboxTheme> {
   manager: CourierInboxThemeManager;
-  unsubscribe: () => void;
 }
 
-export class CourierInboxThemeManager {
+/**
+ * Inbox-specific theme manager that extends the abstract CourierThemeManager.
+ * Provides inbox theme management with light/dark mode support.
+ */
+export class CourierInboxThemeManager extends CourierThemeManager<CourierInboxTheme> {
 
-  // Event IDs
-  private readonly THEME_CHANGE_EVENT: string = 'courier_inbox_theme_change';
-
-  // State
-  private _theme: CourierInboxTheme;
-  private _lightTheme: CourierInboxTheme = defaultLightTheme;
-  private _darkTheme: CourierInboxTheme = defaultDarkTheme;
-  private _target: EventTarget;
-  private _subscriptions: CourierInboxThemeSubscription[] = [];
-
-  // System theme
-  private _userMode: CourierComponentThemeMode;
-  private _systemMode: SystemThemeMode;
-  private _systemThemeCleanup: (() => void) | undefined;
-
-  public setLightTheme(theme: CourierInboxTheme) {
-    this._lightTheme = theme;
-    if (this._systemMode === 'light') {
-      this.updateTheme();
-    }
-  }
-
-  public setDarkTheme(theme: CourierInboxTheme) {
-    this._darkTheme = theme;
-    if (this._systemMode === 'dark') {
-      this.updateTheme();
-    }
-  }
+  // Event ID for inbox theme changes
+  protected readonly THEME_CHANGE_EVENT: string = 'courier_inbox_theme_change';
 
   constructor(initialTheme: CourierInboxTheme) {
-    this._theme = initialTheme;
-    this._target = new EventTarget();
-    this._userMode = 'system' as CourierComponentThemeMode;
-
-    // Get the initial system theme
-    this._systemMode = getSystemThemeMode();
-    this.setLightTheme(defaultLightTheme);
-    this.setDarkTheme(defaultDarkTheme);
-
-    // Set up system theme listener
-    this._systemThemeCleanup = addSystemThemeModeListener((mode: SystemThemeMode) => {
-      this._systemMode = mode;
-      this.updateTheme();
-    });
+    super(initialTheme);
   }
 
   /**
-   * Get the current system theme
+   * Get the default light theme for inbox
    */
-  public get currentSystemTheme(): SystemThemeMode {
-    return this._systemMode;
+  protected getDefaultLightTheme(): CourierInboxTheme {
+    return defaultLightTheme;
   }
 
   /**
-   * Get the current theme
+   * Get the default dark theme for inbox
    */
-  public getTheme() {
-    return this._theme;
+  protected getDefaultDarkTheme(): CourierInboxTheme {
+    return defaultDarkTheme;
   }
 
   /**
-   * Update the theme
+   * Merge the inbox theme with defaults
    */
-  private updateTheme() {
-
-    // Use the user mode or fallback to the system mode
-    const mode = this._userMode === 'system' ? this._systemMode : this._userMode;
-
-    // Get the theme  
-    const theme = mode === 'light' ? this._lightTheme : this._darkTheme;
-
-    // Merge the theme
-    const mergedTheme = mergeTheme(mode, theme);
-
-    // Set the theme
-    this.setTheme(mergedTheme);
+  protected mergeTheme(mode: SystemThemeMode, theme: CourierInboxTheme): CourierInboxTheme {
+    return mergeTheme(mode, theme);
   }
 
   /**
-   * Set the theme and notify all listeners
-   */
-  private setTheme(theme: CourierInboxTheme) {
-    if (theme === this._theme) return;
-    this._theme = theme;
-    this._target.dispatchEvent(new CustomEvent(this.THEME_CHANGE_EVENT, {
-      detail: { theme }
-    }));
-  }
-
-  /**
-   * Set the mode and notify all listeners
-   */
-  public setMode(mode: CourierComponentThemeMode) {
-    this._userMode = mode;
-    this.updateTheme();
-  }
-
-  public get mode(): CourierComponentThemeMode {
-    return this._userMode;
-  }
-
-  /**
-   * Subscribe to theme changes
+   * Subscribe to inbox theme changes
    * @param {Function} callback - Function to run when the theme changes
-   * @returns {CourierInboxThemeSubscription} - Object with remove method to stop listening
+   * @returns {CourierInboxThemeSubscription} - Object with unsubscribe method to stop listening
    */
-  subscribe(callback: (theme: CourierInboxTheme) => void): CourierInboxThemeSubscription {
-    const controller = new AbortController();
-    this._target.addEventListener(this.THEME_CHANGE_EVENT, ((e: Event) => {
-      callback((e as CustomEvent<{ theme: CourierInboxTheme }>).detail.theme);
-    }) as EventListener, { signal: controller.signal });
-
-    const subscription: CourierInboxThemeSubscription = {
-      manager: this,
-      unsubscribe: () => {
-        controller.abort();
-        const index = this._subscriptions.indexOf(subscription);
-        if (index > -1) {
-          this._subscriptions.splice(index, 1);
-        }
-      }
+  public subscribe(callback: (theme: CourierInboxTheme) => void): CourierInboxThemeSubscription {
+    const baseSubscription = super.subscribe(callback);
+    return {
+      ...baseSubscription,
+      manager: this
     };
-
-    this._subscriptions.push(subscription);
-    return subscription;
-  }
-
-  /**
-   * Clean up event listeners
-   */
-  public cleanup() {
-    if (this._systemThemeCleanup) {
-      this._systemThemeCleanup();
-    }
-    this._subscriptions.forEach(subscription => subscription.unsubscribe());
-    this._subscriptions = [];
   }
 
 }
